# Conflicts:
#	pkg/secrets/providers/sops/provider.go
#	pkg/secrets/providers/sops/store.go

+++ features/rbac/jit/notification.go
@@ -543,7 +543,7 @@ func (w *WebhookNotificationService) sendWebhook(ctx context.Context, payload in
 	var lastErr error
 	for attempt := 0; attempt < 3; attempt++ {
 		if attempt > 0 {
-			delay := time.Duration(1<<uint(attempt-1)) * w.retryBase
+			delay := time.Duration(1<<(attempt-1)) * w.retryBase
 			select {
 			case <-ctx.Done():
 				return ctx.Err()
